`zshrc`: split `emacs` configs over to new file; updated aliases.
`e`, which is my most commonly used `emacs` alias, now automatically will start
the daemon if it's not already running & connect to it via `emacsclient`. This
is great for performance, as it means I no longer have to wait for `emacs` to
start up every time I want to edit a file.