`zshrc`: set up `pure` prompt [1] instead of `oh-my-zsh` theme
`pure` is a more minimal prompt that I like more than the `oh-my-zsh` theme I'd
been using. Just added it to my packages list on macOS [2] - TBD on how this'll
be installed on other machines. As I wrote out in explanatory comments, I
considered completely removing my dependency on `oh-my-zsh`, but it seems to be
a bit too useful at the moment with its "niceties", & with this theme removal
it'll be mostly out of the way. Making it even more "out the way" by setting
the update frequency to be longer & the verbosity to be minimal.

[1] https://github.com/sindresorhus/pure
[2] https://github.com/izzygomez/mac-setup/commit/b3c03c8709fddc650464cf5c076e65f0264e9251