use new `zenburn-theme` [1] & auto-remove packages
I was essentially getting an annoying error while using `solarized-emacs`. tl;dr was that it was something like this [2], except ":style" was "unspecified" -- wasn't really interested in figuring out myself, so I decided to just update to use a new theme. went to [3] & discovered [1].

given that I was updating the `my-packages` list with a *deletion*, I added a ChatGPT-created code block to automatically delete packages no longer listed in `my-packages` -- that are not a dependency of something else.

[1] https://github.com/bbatsov/zenburn-emacs
[2] https://github.com/bbatsov/solarized-emacs/issues/447
[3] https://emacsthemes.com/popular/index.html